Migrate Medication to DrugType
