+++ src/main/java/com/rettichlp/unicacityaddon/base/config/ConfigElements.java
@@ -2,7 +2,7 @@ package com.rettichlp.unicacityaddon.base.config;
 
 import com.rettichlp.unicacityaddon.UnicacityAddon;
 import com.rettichlp.unicacityaddon.base.enums.faction.Faction;
-import com.rettichlp.unicacityaddon.base.enums.faction.badfaction.DrugPurity;
+import com.rettichlp.unicacityaddon.base.enums.faction.DrugPurity;
 import com.rettichlp.unicacityaddon.base.text.ColorCode;
 import com.rettichlp.unicacityaddon.base.text.Message;
 import joptsimple.internal.Strings;
